requirements.txt
python-telegram-bot==21.6
pymongo==4.6.1
APScheduler==3.10.4
python-dotenv==1.0.1
